+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 - added support for unmodifiable test-provided parameters via `TestConfigs.fixed(..)`
+- extended JSON response
 
 ## [0.2.2] - 2019-03-28
 - fix artifact deployment to sonatype